+++ packages/react/Radio/src/components/Skeleton/skeleton.types.ts
@@ -1,0 +1,7 @@
+import { SkeletonProps as SkeletonBaseProps } from "@nimbus-ds/skeleton";
+import { RadioProps } from "../../radio.types";
+
+export type SkeletonProps = Partial<Pick<SkeletonBaseProps, "width">> &
+  Partial<Pick<RadioProps, "as">> & {
+    "data-testid"?: string;
+  };