+++ packages/styles/src/index.ts
@@ -14,6 +14,7 @@ export { text } from "./packages/text";
 export { title } from "./packages/title";
 export { toast } from "./packages/toast";
 export { list } from "./packages/list";
+export { radio } from "./packages/radio";
 export { toggle } from "./packages/toggle";
 export { link } from "./packages/link";
 export { tooltip } from "./packages/tooltip";