+++ packages/styles/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Nimbus Styles deprive all styles needed to build nimbus components.
 
+## 2022-10-05
+
+### 🎉 New features
+
+- Added new style pack for radio component. ([#36](https://github.com/TiendaNube/nimbus-design-system/pull/36) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2022-10-03
 
 ### 🎉 New features